pyproject.toml
[tool.poetry]
name = "qol-backend-api"
version = "0.1.0"
description = "Quest of Life Backend API"
authors = ["Your Name <your@email.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.0"
aiomysql = "^0.2.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ipython = "^9.3.0"
pylint = "^3.0.0"
mypy = "^1.8.0"
ruff = "^0.2.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 